Armv8: flush_cache: Use equal-width data type
Use an equal-width datatype for the cacheline
sizes so that the start address is not
loosing its upper 32bit part in the rounding
down operation.

+++ src/kern/arm/64/mmu-arm-64.cpp
@@ -26,7 +26,7 @@ template< unsigned long Flush_area, bool Ram >
 void Mmu<Flush_area, Ram>::flush_cache(void const *start,
 				       void const *end)
 {
-  unsigned i = icache_line_size(), d = dcache_line_size();
+  unsigned long i = icache_line_size(), d = dcache_line_size();
   __asm__ __volatile__ (
       "1:  dc civac, %[i]  \n"
       "    ic ivau,  %[i]  \n"
